Distributing wallet responsabilities among brokers, avoiding so many ifs and elses
