+++ federator/config/mumbai.json
@@ -1,11 +1,11 @@
 {
     "name": "mumbai",
-    "bridge": "0xf2394e25954713a25b4e5a3d3b5a57b313b84a6b",
-    "federation": "0x4cd634d1399eeeb61294bdf9a2df7588f7d4af8f",
-    "multiSig": "0x7cfded4ccdac46d2d1003b4bc9058b2135feb79e",
-    "allowTokens": "0x458ae8c887440e89251d6be560d11802aad606fd",
+    "bridge": "0x554b63f926a8120242c10e69a61fff46abeda07d",
+    "federation": "0x933a5049dea3ba93aeb0a79f516e83d6ad206b8a",
+    "multiSig": "0xab2a1321d4a049d67bdb8ac05cc3e036606b9383",
+    "allowTokens": "0xbf52a73032039f274155e2467aa31de6ef1310f5",
     "chainId": 80001,
-    "testToken": "0x9956d17a0615e2af9f3745c87a55db05fcc50329",
+    "testToken": "0x995491ebafe3e286201bf22af1af9803b5c5d9b5",
     "host": "https://polygon-mumbai.infura.io/v3/399500b5679b442eb991fefee1c5bfdc",
-    "fromBlock": 43918256
+    "fromBlock": 44851963
 }
