+++ federator/src/types/HathorResponseTypes.ts
@@ -1,0 +1,13 @@
+import { Data } from './hathorEvent';
+
+export type HathorResponse = {
+  success: boolean;
+  message?: string;
+  error?: string;
+  errorCode?: string;
+};
+export type CreateProposalResponse = HathorResponse & { txHex: string };
+export type GetAddressResponse = HathorResponse & { address: string };
+export type GetMySignatureResponse = HathorResponse & { signatures: string };
+export type StatusResponse = HathorResponse & { statusCode: number; statusMessage: string };
+export type DecodeResponse = HathorResponse & { tx: Data };
